+++ repeat/api/tests/test_util.py
@@ -1,0 +1,9 @@
+import django.test
+
+from api import util
+
+# TODO
+# class PDFToTextTests(django.test.TestCase):
+
+#     def test_pdf_to_text(self):
+#         self.assertEqual("", util.pdf_to_text("TODO"))
